feat(test-runner): move environment.js and config.js to test-runner (#2052)
* feat(test-runner): move environment.js and config.js to test-runner

NODE-2041
